+++ tests/testthat/test-convertPedigree.R
@@ -1,7 +1,7 @@
 test_that("ped2add produces correct matrix dims, values, and dimnames for hazard", {
   tolerance <- 1e-10
   data(hazard)
-  add <- ped2add(hazard)
+  add <- ped2add(hazard,sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(hazard), nrow(hazard)))
   # Check several values
@@ -21,7 +21,7 @@ test_that("ped2add produces correct matrix dims, values, and dimnames for hazard
 test_that("ped2add produces correct matrix dims, values, and dimnames for alternative transpose", {
   tolerance <- 1e-10
   data(hazard)
-  add <- ped2add(hazard, tcross.alt.crossprod = TRUE)
+  add <- ped2add(hazard, tcross.alt.crossprod = TRUE,sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(hazard), nrow(hazard)), tolerance = tolerance)
   # Check several values
@@ -42,7 +42,7 @@ test_that("ped2add produces correct matrix dims, values, and dimnames for altern
 test_that("ped2add produces correct matrix dims, values, and dimnames for inbreeding data", {
   tolerance <- 1e-10
   data(inbreeding)
-  add <- ped2add(inbreeding)
+  add <- ped2add(inbreeding,sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(inbreeding), nrow(inbreeding)), tolerance = tolerance)
   # Check several values
@@ -62,7 +62,7 @@ test_that("ped2add produces correct matrix dims, values, and dimnames for inbree
 test_that("ped2add produces correct matrix dims, values, and dimnames for inbreeding data with alternative transpose", {
   tolerance <- 1e-10
   data(inbreeding)
-  add <- ped2add(inbreeding, transpose_method = "star")
+  add <- ped2add(inbreeding, transpose_method = "star",sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(inbreeding), nrow(inbreeding)))
   # Check several values
@@ -81,7 +81,7 @@ test_that("ped2add produces correct matrix dims, values, and dimnames for inbree
 test_that("ped2add produces correct matrix dims, values, and dimnames for inbreeding data with 2nd alternative transpose", {
   tolerance <- 1e-10
   data(inbreeding)
-  add <- ped2add(inbreeding, transpose_method = "crossprod")
+  add <- ped2add(inbreeding, transpose_method = "crossprod",sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(inbreeding), nrow(inbreeding)))
   # Check several values
@@ -101,7 +101,7 @@ test_that("ped2add produces correct matrix dims, values, and dimnames for inbree
 test_that("ped2add flattens diagonal for inbreeding data", {
   tolerance <- 1e-10
   data(inbreeding)
-  add <- ped2add(inbreeding, flatten.diag = TRUE)
+  add <- ped2add(inbreeding, flatten.diag = TRUE,sparse = FALSE)
   # Check dimension
   expect_equal(dim(add), c(nrow(inbreeding), nrow(inbreeding)), tolerance = tolerance)
   # Check several values
@@ -121,7 +121,7 @@ test_that("ped2mit produces correct matrix dims, values, and dimnames for inbree
   tolerance <- 1e-10
   # Check dimension
   data(inbreeding)
-  mit <- ped2mit(inbreeding)
+  mit <- ped2mit(inbreeding,sparse = FALSE)
   # Check dimension
   expect_equal(dim(mit), c(nrow(inbreeding), nrow(inbreeding)))
   # Check several values
@@ -142,7 +142,7 @@ test_that("ped2mit produces correct matrix dims, values, and dimnames for inbree
   tolerance <- 1e-10
   # Check dimension
   data(inbreeding)
-  mit <- ped2mit(inbreeding)
+  mit <- ped2mit(inbreeding,sparse = FALSE)
   # Check dimension
   expect_equal(dim(mit), c(nrow(inbreeding), nrow(inbreeding)), tolerance = tolerance)
   # Check several values
@@ -164,7 +164,7 @@ test_that("ped2cn produces correct matrix dims, values, and dimnames", {
 
   # Check dimension
   data(inbreeding)
-  cn <- ped2cn(inbreeding)
+  cn <- ped2cn(inbreeding,sparse = FALSE)
   expect_equal(dim(cn), c(
     nrow(inbreeding),
     nrow(inbreeding)
@@ -189,7 +189,7 @@ test_that("ped2cn produces correct matrix dims, values, and dimnames", {
 test_that("ped2ce produces correct matrix dims, values, and dimnames", {
   tolerance <- 1e-10
   data(inbreeding)
-  ce <- ped2ce(inbreeding)
+  ce <- ped2ce(inbreeding,sparse = FALSE)
   expect_equal(dim(ce), c(nrow(inbreeding), nrow(inbreeding)), tolerance = tolerance)
   # Check several values
   # expect_true(all(diag(ce) == 1))
@@ -207,7 +207,7 @@ test_that("ped2ce produces correct matrix dims, values, and dimnames", {
 
 test_that("ped2add verbose prints updates", {
   data(hazard)
-  expect_output(ped2add(hazard, verbose = TRUE), regexp = "Family Size =")
+  expect_output(ped2add(hazard, verbose = TRUE,sparse = FALSE), regexp = "Family Size =")
 })